+++ packages/components/src/components/propTypes/index.ts
@@ -72,6 +72,7 @@ import type { FileCardListProps } from "@/components/FileCardList";
 import type { AccentBoxProps } from "@/components/AccentBox";
 import type { ColumnLayoutProps } from "@/components/ColumnLayout";
 import type { MenuTriggerProps } from "@/components/OverlayTrigger";
+import type { ProgressBarProps } from "@/components/ProgressBar";
 
 export * from "./types";
 
@@ -129,6 +130,7 @@ export interface FlowComponentPropsTypes {
   Popover: PopoverProps;
   PopoverTrigger: PopoverTriggerProps;
   PasswordCreationField: PasswordCreationFieldProps;
+  ProgressBar: ProgressBarProps;
   Radio: RadioProps;
   RadioButton: RadioButtonProps;
   RadioGroup: RadioGroupProps;
@@ -205,6 +207,7 @@ const propsContextSupportingComponentsMap: Record<
   Popover: true,
   PopoverTrigger: true,
   PasswordCreationField: true,
+  ProgressBar: true,
   RadioButton: true,
   RadioGroup: true,
   SearchField: true,