+++ packages/components/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-alpha.185](https://github.com/mittwald/flow/compare/0.2.0-alpha.184...0.2.0-alpha.185) (2025-05-15)
+
+### Features
+
+* **AreaChart:** implement emptyView for empty/error display ([#1463](https://github.com/mittwald/flow/issues/1463)) ([7f5c8ca](https://github.com/mittwald/flow/commit/7f5c8ca2db56578b2cbc2f974e50636c86733504))
+
+# [0.2.0-alpha.184](https://github.com/mittwald/flow/compare/0.2.0-alpha.183...0.2.0-alpha.184) (2025-05-14)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# [0.2.0-alpha.183](https://github.com/mittwald/flow/compare/0.2.0-alpha.182...0.2.0-alpha.183) (2025-05-14)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# [0.2.0-alpha.182](https://github.com/mittwald/flow/compare/0.2.0-alpha.181...0.2.0-alpha.182) (2025-05-14)
+
+### Bug Fixes
+
+* **List:** update tile translation ([097dc27](https://github.com/mittwald/flow/commit/097dc27d968f975df613dd27e26e7730df21e36b))
+
+### Features
+
+* **IllustratedMessage:** support progress bar in illustrated message ([8819fe0](https://github.com/mittwald/flow/commit/8819fe088f99981240906d905b70389b9096de10))
+* **Text:** support icon in text ([6a847ca](https://github.com/mittwald/flow/commit/6a847cab8f17c9688d9b94422e63bf02437f63fb))
+
+# [0.2.0-alpha.181](https://github.com/mittwald/flow/compare/0.2.0-alpha.180...0.2.0-alpha.181) (2025-05-13)
+
+### Bug Fixes
+
+* **ContextMenu:** support default selected keys in context menu section ([#1455](https://github.com/mittwald/flow/issues/1455)) ([4753d36](https://github.com/mittwald/flow/commit/4753d365def874192da28a6e85b80cd410fd34ce))
+
+# [0.2.0-alpha.180](https://github.com/mittwald/flow/compare/0.2.0-alpha.179...0.2.0-alpha.180) (2025-05-12)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
 # [0.2.0-alpha.179](https://github.com/mittwald/flow/compare/0.2.0-alpha.178...0.2.0-alpha.179) (2025-05-12)
 
 **Note:** Version bump only for package @mittwald/flow-react-components